Revert "debian/rules: Set Distribution: Field"
This reverts commit b88f0269a6609521cb6f4fbe8e4c000978430b55.

The intent of that commit was to set the Distribution in the control
file to the same as the build environment, easing the maintenance of
multiple apt repositories.
Since debian/rules now strips "minimum version" strings from the
dependency information, thus allowing a single build to have satisfiable
dependencies on a wider range of install environments, it is now
redundant.