push_to_ppa.yml: switch to -S instead of --build=source
I prefer using longoptions in scripts, they better describe the intent
but sadly not an option here...

Both -S and --build=source are available to dpkg-buildpackage, but
the latter is not available in debuild.
Therefore --build=source is passed to dpkg-buildpackage, which
results in debuild later expecting *_${arch}.changes file, which does not
exist.
Using -S resolves this, since gbp passes it to debuild which in turn
passes it to dpkg-buildpackage.

At least that is what I think is happening.