weekly_ppa.yml: add weekly checkout
depth=1 is not going to be all that practical so instead of fetching the
full git history, only get since last -alpha tag.
This is so "git describe" can output something which can be used as a
debianised version string and that gbp can use to create the .orig.tar

I other words, the debian version string is a direct pointer to the
commit in git.