d/control: Remove curl, ca-certificates from Build-Depends
These are redundent, I was using them to have
git-pbuilder download the static-fluidsynth for
Bionic.

The Launchpad PPA builders don't have network, so
I have the github workflow scripts download and
bundle the tarball in the debian.tar

The cmake is also redundent on newer than Bionic
but leaving it there for now.

Gbp-Dch: Ignore